Overhaul UI selection dialog
Removed the ability to create/rename files/folders within the dialog. Improved the readability of the function and simplified the filename/path handling as prepareForRead wants a string array of absolute paths rather than a split into file and path. Updated prepareForRead to reflect this modification. Overall behaviour in both functions is unchanged.