Update internal input validation
Removed the need to pass an empty filename variable when operating a cloud-based search. Added check to ensure the given radarIDs are valid sites. This is done by attempting to convert the passed string array into the radarID enumeration class. This will error if one or more of the ID are not valid. Added Output argument validation to io functions for conformity.